@@ -1,15 +1,14 @@
 module github.com/itchio/hush
 
-go 1.13
+go 1.21
 
 require (
 	github.com/BurntSushi/toml v0.3.1
 	github.com/dchest/safefile v0.0.0-20151022103144-855e8d98f185
 	github.com/itchio/boar v0.0.0-20200305195307-d2befc01fa9e
-	github.com/itchio/dmcunrar-go v0.0.0-20200303200038-d0b2e3ba28f6 // indirect
-	github.com/itchio/go-itchio v0.0.0-20200301160906-508e37e43b9c
+	github.com/itchio/go-itchio v0.0.0-20251229221754-554b6b9748f0
 	github.com/itchio/headway v0.0.0-20200301160421-e15721f23905
-	github.com/itchio/httpkit v0.0.0-20200304092139-56c2e1e88c9b
+	github.com/itchio/httpkit v0.0.0-20251229220831-fe361f261a25
 	github.com/itchio/intact v0.0.0-20200301161822-f8c4a3336c2a
 	github.com/itchio/lake v0.0.0-20200305150023-cc4284ec2b2a
 	github.com/itchio/ox v0.0.0-20200301160301-4e131878ba64
@@ -17,6 +16,48 @@ require (
 	github.com/itchio/screw v0.0.0-20200301160148-75fc2d65fb38
 	github.com/mitchellh/mapstructure v1.1.2
 	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.5.1
-	golang.org/x/sys v0.0.0-20200302150141-5c8b2ff67527 // indirect
+	github.com/stretchr/testify v1.11.1
+)
+
+require (
+	github.com/basgys/goxml2json v1.1.0 // indirect
+	github.com/certifi/gocertifi v0.0.0-20200211180108-c7c1fbc02894 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/detailyang/go-fallocate v0.0.0-20180908115635-432fa640bd2e // indirect
+	github.com/efarrer/iothrottler v0.0.1 // indirect
+	github.com/getlantern/context v0.0.0-20190109183933-c447772a6520 // indirect
+	github.com/getlantern/errors v1.0.1 // indirect
+	github.com/getlantern/golog v0.0.0-20230503153817-8e72de7e0a65 // indirect
+	github.com/getlantern/hex v0.0.0-20190417191902-c6586a6fe0b7 // indirect
+	github.com/getlantern/hidden v0.0.0-20190325191715-f02dbb02be55 // indirect
+	github.com/getlantern/idletiming v0.0.0-20231030193830-6767b09f86db // indirect
+	github.com/getlantern/iptool v0.0.0-20210721034953-519bf8ce0147 // indirect
+	github.com/getlantern/mtime v0.0.0-20200417132445-23682092d1f7 // indirect
+	github.com/getlantern/netx v0.0.0-20211206143627-7ccfeb739cbd // indirect
+	github.com/getlantern/ops v0.0.0-20200403153110-8476b16edcd6 // indirect
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/gogs/chardet v0.0.0-20191104214054-4b6791f73a28 // indirect
+	github.com/golang/protobuf v1.3.4 // indirect
+	github.com/itchio/arkive v0.0.0-20200301155608-aeded25a0494 // indirect
+	github.com/itchio/dash v0.0.0-20200305193527-478357e813a2 // indirect
+	github.com/itchio/dmcunrar-go v0.0.0-20241212230744-bd86d6b265e9 // indirect
+	github.com/itchio/kompress v0.0.0-20200301155538-5c2eecce9e51 // indirect
+	github.com/itchio/pelican v0.0.0-20200305192721-308c3772290a // indirect
+	github.com/itchio/sevenzip-go v0.0.0-20251211223907-deaa375308d8 // indirect
+	github.com/itchio/spellbook v0.0.0-20200301161431-a4bdbea6b725 // indirect
+	github.com/itchio/wizardry v0.0.0-20200301161332-e8c8c4a5a488 // indirect
+	github.com/klauspost/compress v1.10.2 // indirect
+	github.com/mitchellh/copystructure v1.0.0 // indirect
+	github.com/mitchellh/reflectwalk v1.0.1 // indirect
+	github.com/nightlyone/lockfile v0.0.0-20200124072040-edb130adc195 // indirect
+	github.com/oxtoacart/bpool v0.0.0-20190530202638-03653db5a59c // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
+	go.uber.org/multierr v1.6.0 // indirect
+	go.uber.org/zap v1.19.1 // indirect
+	golang.org/x/net v0.35.0 // indirect
+	golang.org/x/sys v0.30.0 // indirect
+	golang.org/x/text v0.22.0 // indirect
+	golang.org/x/time v0.0.0-20191024005414-555d28b269f0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
